Add a Setting and SettingGroup concept to the Plugin class.
This is expected to replace the custom OptionParser thing we've been
using so far.

Of particular interest is the fact that the VM object itself does not
get these attributes set. Instead, the VM object gets a ._config
attribute which is a dict that maps option names to Setting objects. The
Setting objects then holds the values. This will require changes across
all of VMBuilder.

Tests are included.